Move student submission service class to service package
